requirements.txt
pydantic
vchat
aiohttp
openai
qq-botpy
chardet~=5.1.0
Pillow
beautifulsoup4
googlesearch-python
readability-lxml
quart
psutil
lxml_html_clean
colorlog
aiocqhttp
pyjwt
apscheduler